requirements.txt
Django==5.2

python-decouple==3.8

djangorestframework==3.16.0

djangorestframework-simplejwt==5.5.0

PyJWT==2.9.0

django-ratelimit==4.1.0